fix(comments): prevent comment from collapsing if child is focused with link hash
